Revert "rename types.py"
This reverts commit d1ebc0a097b8259f422470d8a549ccc82bdb0df9.
